+++ src/shortest-path/indexed-brandes.d.ts
@@ -1,9 +1,10 @@
-import Graph from 'graphology-types';
+import Graph, {Attributes} from 'graphology-types';
+import {MinimalEdgeMapper} from 'graphology-utils/getters';
 import FixedStack from 'mnemonist/fixed-stack';
 import {
-  OutboundNeighborhoodIndex,
-  WeightedOutboundNeighborhoodIndex
-} from 'graphology-indices/neighborhood/outbound';
+  NeighborhoodIndex,
+  WeightedNeighborhoodIndex
+} from 'graphology-indices/neighborhood';
 
 type IndexedBrandesResult = [
   FixedStack<number>,
@@ -15,12 +16,20 @@ type IndexedBrandesFunction = (sourceIndex: number) => IndexedBrandesResult;
 
 interface ICreateUnweightedIndexedBrandes {
   (graph: Graph): IndexedBrandesFunction;
-  index: OutboundNeighborhoodIndex;
+  index: NeighborhoodIndex;
 }
 
-interface ICreateDijkstraIndexedBrandes {
-  (graph: Graph, weightAttribute?: string): IndexedBrandesFunction;
-  index: WeightedOutboundNeighborhoodIndex;
+interface ICreateDijkstraIndexedBrandes<
+  NodeAttributes extends Attributes = Attributes,
+  EdgeAttributes extends Attributes = Attributes
+> {
+  (
+    graph: Graph<NodeAttributes, EdgeAttributes>,
+    getEdgeWeight?:
+      | keyof EdgeAttributes
+      | MinimalEdgeMapper<number, EdgeAttributes>
+  ): IndexedBrandesFunction;
+  index: WeightedNeighborhoodIndex;
 }
 
 declare const createUnweightedIndexedBrandes: ICreateUnweightedIndexedBrandes;